Haptics fixed

+++ app/src/main/jni/Doom3Quest/VrInput.h
@@ -33,4 +33,4 @@ void IN_VRInputFrame( engine_t* engine );
 uint32_t IN_VRGetButtonState( int controllerIndex );
 XrVector2f IN_VRGetJoystickState( int controllerIndex );
 XrPosef IN_VRGetPose( int controllerIndex );
-void INVR_Vibrate( int duration, int chan, float intensity );
+void INVR_Vibrate( float duration, int chan, float intensity );
